src/requirements.txt
python-dotenv>=0.20.0
APScheduler>=3.9.1
pyaml-env>=1.1.5
requests>=2.27.1
SQLAlchemy>=1.4
psycopg2>=2.9.3